+++ src/cucumber.rs
@@ -143,8 +143,8 @@ where
     ///
     /// Output with a regular [`Cucumber::run()`]:
     /// <script
-    ///     id="asciicast-hMyH3IYbHRFXT1yf84tXDNl2r"
-    ///     src="https://asciinema.org/a/hMyH3IYbHRFXT1yf84tXDNl2r.js"
+    ///     id="asciicast-0d92qlT8Mbc4WXyvRbHJmjsqN"
+    ///     src="https://asciinema.org/a/0d92qlT8Mbc4WXyvRbHJmjsqN.js"
     ///     async data-autoplay="true" data-rows="17">
     /// </script>
     ///
@@ -178,8 +178,8 @@ where
     /// # futures::executor::block_on(fut);
     /// ```
     /// <script
-    ///     id="asciicast-UsaG9kMnn40nW8y4vcmXOE2tT"
-    ///     src="https://asciinema.org/a/UsaG9kMnn40nW8y4vcmXOE2tT.js"
+    ///     id="asciicast-IHLxMEgku9BtBVkR4k2DtOjMd"
+    ///     src="https://asciinema.org/a/IHLxMEgku9BtBVkR4k2DtOjMd.js"
     ///     async data-autoplay="true" data-rows="21">
     /// </script>
     ///
@@ -223,8 +223,8 @@ where
     ///
     /// Output with a regular [`Cucumber::run()`]:
     /// <script
-    ///     id="asciicast-hMyH3IYbHRFXT1yf84tXDNl2r"
-    ///     src="https://asciinema.org/a/hMyH3IYbHRFXT1yf84tXDNl2r.js"
+    ///     id="asciicast-0d92qlT8Mbc4WXyvRbHJmjsqN"
+    ///     src="https://asciinema.org/a/0d92qlT8Mbc4WXyvRbHJmjsqN.js"
     ///     async data-autoplay="true" data-rows="17">
     /// </script>
     ///
@@ -234,7 +234,6 @@ where
     /// # use std::convert::Infallible;
     /// #
     /// # use async_trait::async_trait;
-    /// # use futures::FutureExt as _;
     /// # use cucumber::WorldInit;
     /// #
     /// # #[derive(Debug, WorldInit)]
@@ -272,8 +271,8 @@ where
     ///     Then the dog is not hungry
     /// ```
     /// <script
-    ///     id="asciicast-UsaG9kMnn40nW8y4vcmXOE2tT"
-    ///     src="https://asciinema.org/a/UsaG9kMnn40nW8y4vcmXOE2tT.js"
+    ///     id="asciicast-IHLxMEgku9BtBVkR4k2DtOjMd"
+    ///     src="https://asciinema.org/a/IHLxMEgku9BtBVkR4k2DtOjMd.js"
     ///     async data-autoplay="true" data-rows="21">
     /// </script>
     ///
@@ -323,8 +322,8 @@ where
     ///
     /// Output with a regular [`Cucumber::run()`]:
     /// <script
-    ///     id="asciicast-hMyH3IYbHRFXT1yf84tXDNl2r"
-    ///     src="https://asciinema.org/a/hMyH3IYbHRFXT1yf84tXDNl2r.js"
+    ///     id="asciicast-0d92qlT8Mbc4WXyvRbHJmjsqN"
+    ///     src="https://asciinema.org/a/0d92qlT8Mbc4WXyvRbHJmjsqN.js"
     ///     async data-autoplay="true" data-rows="17">
     /// </script>
     ///
@@ -358,8 +357,8 @@ where
     /// # futures::executor::block_on(fut);
     /// ```
     /// <script
-    ///     id="asciicast-BD1mPjYGELD6oWNKW8lTlyvDR"
-    ///     src="https://asciinema.org/a/BD1mPjYGELD6oWNKW8lTlyvDR.js"
+    ///     id="asciicast-ox14HynkBIw8atpfhyfvKrsO3"
+    ///     src="https://asciinema.org/a/ox14HynkBIw8atpfhyfvKrsO3.js"
     ///     async data-autoplay="true" data-rows="19">
     /// </script>
     ///
@@ -410,8 +409,8 @@ where
     /// # futures::executor::block_on(fut);
     /// ```
     /// <script
-    ///     id="asciicast-mDDqxWHzUaK19P0L2R2g4XRp2"
-    ///     src="https://asciinema.org/a/mDDqxWHzUaK19P0L2R2g4XRp2.js"
+    ///     id="asciicast-UcipuopO6IFEsIDty6vaJlCH9"
+    ///     src="https://asciinema.org/a/UcipuopO6IFEsIDty6vaJlCH9.js"
     ///     async data-autoplay="true" data-rows="21">
     /// </script>
     ///
@@ -446,8 +445,8 @@ where
     /// # futures::executor::block_on(fut);
     /// ```
     /// <script
-    ///     id="asciicast-qKp8Hevrb6732mMUT7VduvxJc"
-    ///     src="https://asciinema.org/a/qKp8Hevrb6732mMUT7VduvxJc.js"
+    ///     id="asciicast-ofOljvyEMb41OTLhE081QKv68"
+    ///     src="https://asciinema.org/a/ofOljvyEMb41OTLhE081QKv68.js"
     ///     async data-autoplay="true" data-rows="24">
     /// </script>
     ///
@@ -506,8 +505,8 @@ where
     /// # futures::executor::block_on(fut);
     /// ```
     /// <script
-    ///     id="asciicast-mDDqxWHzUaK19P0L2R2g4XRp2"
-    ///     src="https://asciinema.org/a/mDDqxWHzUaK19P0L2R2g4XRp2.js"
+    ///     id="asciicast-UcipuopO6IFEsIDty6vaJlCH9"
+    ///     src="https://asciinema.org/a/UcipuopO6IFEsIDty6vaJlCH9.js"
     ///     async data-autoplay="true" data-rows="21">
     /// </script>
     ///
@@ -564,8 +563,8 @@ where
     /// # futures::executor::block_on(fut);
     /// ```
     /// <script
-    ///     id="asciicast-qKp8Hevrb6732mMUT7VduvxJc"
-    ///     src="https://asciinema.org/a/qKp8Hevrb6732mMUT7VduvxJc.js"
+    ///     id="asciicast-ofOljvyEMb41OTLhE081QKv68"
+    ///     src="https://asciinema.org/a/ofOljvyEMb41OTLhE081QKv68.js"
     ///     async data-autoplay="true" data-rows="24">
     /// </script>
     ///
@@ -666,8 +665,8 @@ where
     ///     Then the dog is not hungry
     /// ```
     /// <script
-    ///     id="asciicast-WbP3PIQR5M7Iznd7uLnjg2ytr"
-    ///     src="https://asciinema.org/a/WbP3PIQR5M7Iznd7uLnjg2ytr.js"
+    ///     id="asciicast-0KvTxnfaMRjsvsIKsalS611Ta"
+    ///     src="https://asciinema.org/a/0KvTxnfaMRjsvsIKsalS611Ta.js"
     ///     async data-autoplay="true" data-rows="14">
     /// </script>
     ///
@@ -1008,8 +1007,8 @@ where
     ///     Then the dog is not hungry
     /// ```
     /// <script
-    ///     id="asciicast-WbP3PIQR5M7Iznd7uLnjg2ytr"
-    ///     src="https://asciinema.org/a/WbP3PIQR5M7Iznd7uLnjg2ytr.js"
+    ///     id="asciicast-0KvTxnfaMRjsvsIKsalS611Ta"
+    ///     src="https://asciinema.org/a/0KvTxnfaMRjsvsIKsalS611Ta.js"
     ///     async data-autoplay="true" data-rows="14">
     /// </script>
     ///